requirements.txt
flask==3.1.0
pyserial==3.5
openai==1.82.0
apscheduler==3.11.0
requests==2.32.3
python-dotenv==1.1.0
